# web/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerCreate": {
        "additionalProperties": false,
        "properties": {
          "method": {
            "$ref": "#/components/schemas/AnswerMethod",
            "default": "TEXT"
          },
          "question_formulation_id": {
            "format": "uuid",
            "title": "Question Formulation Id",
            "type": "string"
          },
          "text": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "question_formulation_id",
          "text"
        ],
        "title": "AnswerCreate",
        "type": "object"
      },
      "AnswerDetail": {
        "properties": {
          "answer_id": {
            "format": "uuid",
            "title": "Answer Id",
            "type": "string"
          },
          "consolidation_round": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consolidation Round"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "evaluations": {
            "items": {
              "$ref": "#/components/schemas/EvaluationRead"
            },
            "title": "Evaluations",
            "type": "array"
          },
          "final_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "hint_used": {
            "default": false,
            "title": "Hint Used",
            "type": "boolean"
          },
          "intent": {
            "$ref": "#/components/schemas/SessionIntent"
          },
          "learning_item_id": {
            "format": "uuid",
            "title": "Learning Item Id",
            "type": "string"
          },
          "needs_self_grade": {
            "title": "Needs Self Grade",
            "type": "boolean"
          },
          "overridden_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overridden At"
          },
          "override_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Note"
          },
          "override_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "question_formulation_id": {
            "format": "uuid",
            "title": "Question Formulation Id",
            "type": "string"
          },
          "reference": {
            "$ref": "#/components/schemas/Reference"
          },
          "resolved_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolver_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolver Version"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleChange"
              },
              {
                "type": "null"
              }
            ]
          },
          "session": {
            "$ref": "#/components/schemas/SessionRead"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "xp": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XpResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "answer_id",
          "learning_item_id",
          "question_formulation_id",
          "intent",
          "text",
          "evaluation",
          "resolved_outcome",
          "resolver_version",
          "override_outcome",
          "final_outcome",
          "needs_self_grade",
          "schedule",
          "reference",
          "session",
          "evaluations",
          "override_note",
          "overridden_at",
          "created_at"
        ],
        "title": "AnswerDetail",
        "type": "object"
      },
...
      "AnswerResult": {
        "properties": {
          "answer_id": {
            "format": "uuid",
            "title": "Answer Id",
            "type": "string"
          },
          "consolidation_round": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consolidation Round"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "hint_used": {
            "default": false,
            "title": "Hint Used",
            "type": "boolean"
          },
          "intent": {
            "$ref": "#/components/schemas/SessionIntent"
          },
          "learning_item_id": {
            "format": "uuid",
            "title": "Learning Item Id",
            "type": "string"
          },
          "needs_self_grade": {
            "title": "Needs Self Grade",
            "type": "boolean"
          },
          "override_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "question_formulation_id": {
            "format": "uuid",
            "title": "Question Formulation Id",
            "type": "string"
          },
          "reference": {
            "$ref": "#/components/schemas/Reference"
          },
          "resolved_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "resolver_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolver Version"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleChange"
              },
              {
                "type": "null"
              }
            ]
          },
          "session": {
            "$ref": "#/components/schemas/SessionRead"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "xp": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XpResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "answer_id",
          "learning_item_id",
          "question_formulation_id",
          "intent",
          "text",
          "evaluation",
          "resolved_outcome",
          "resolver_version",
          "override_outcome",
          "final_outcome",
          "needs_self_grade",
          "schedule",
          "reference",
          "session"
        ],
        "title": "AnswerResult",
        "type": "object"
      },
...
      "ConsolidationPlan": {
        "description": "What \"I have studied this concept\" would start, shown before starting (items, rounds,\nanswers), or the unfinished batch it would resume.",
        "properties": {
          "answers_in_batch": {
            "title": "Answers In Batch",
            "type": "integer"
          },
          "batch_items": {
            "title": "Batch Items",
            "type": "integer"
          },
          "concept_active": {
            "title": "Concept Active",
            "type": "boolean"
          },
          "concept_id": {
            "format": "uuid",
            "title": "Concept Id",
            "type": "string"
          },
          "new_items": {
            "title": "New Items",
            "type": "integer"
          },
          "remaining_after_batch": {
            "title": "Remaining After Batch",
            "type": "integer"
          },
          "rounds_per_item": {
            "title": "Rounds Per Item",
            "type": "integer"
          },
          "unfinished": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "unfinished_items": {
            "title": "Unfinished Items",
            "type": "integer"
          }
        },
        "required": [
          "concept_id",
          "concept_active",
          "unfinished",
          "unfinished_items",
          "new_items",
          "batch_items",
          "rounds_per_item",
          "answers_in_batch",
          "remaining_after_batch"
        ],
        "title": "ConsolidationPlan",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HintState": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "revealed": {
            "title": "Revealed",
            "type": "boolean"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "available",
          "revealed",
          "text"
        ],
        "title": "HintState",
        "type": "object"
      },
...
      "OverrideCreate": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "outcome": {
            "$ref": "#/components/schemas/ReviewOutcome"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "OverrideCreate",
        "type": "object"
      },
...
      "ReviewOutcome": {
        "description": "docs/PROJECT_SPEC.md \u00a746. Decided by the ReviewOutcomeResolver (or the user's override),\nnever by the AI directly.",
        "enum": [
          "AGAIN",
          "HARD",
          "GOOD",
          "EASY"
        ],
        "title": "ReviewOutcome",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Adaptive AI Learning Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/answers/{answer_id}": {
      "get": {
        "operationId": "get_answer_api_v1_answers__answer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "answer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Answer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Answer",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/answers/{answer_id}/evaluate": {
      "post": {
        "description": "Evaluates again after a failed (or unconfigured) attempt.",
        "operationId": "retry_evaluation_api_v1_answers__answer_id__evaluate_post",
        "parameters": [
          {
            "in": "path",
            "name": "answer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Answer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Evaluation",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/answers/{answer_id}/override": {
      "post": {
        "description": "The user's own grade. Keeps the AI evaluation; see docs/API.md for replay rules.",
        "operationId": "override_api_v1_answers__answer_id__override_post",
        "parameters": [
          {
            "in": "path",
            "name": "answer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Answer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Override",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/concepts/{concept_id}/consolidation": {
      "get": {
        "description": "What \"I have studied this concept\" would do: resume an unfinished batch, or start the\nnext batch (item count, rounds and answers, shown before starting).",
        "operationId": "consolidation_plan_api_v1_concepts__concept_id__consolidation_get",
        "parameters": [
          {
            "in": "path",
            "name": "concept_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Concept Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsolidationPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Consolidation Plan",
        "tags": [
          "review"
        ]
      },
      "post": {
        "description": "\"I have studied this concept\": resumes the unfinished batch, or starts the next one\n(each item asked three times in a row). 409 `nothing_to_consolidate` when no NEW item is\nwaiting. Answer it through the usual review-session endpoints.",
        "operationId": "start_consolidation_api_v1_concepts__concept_id__consolidation_post",
        "parameters": [
          {
            "in": "path",
            "name": "concept_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Concept Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Consolidation",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/review-sessions": {
      "post": {
        "description": "Builds the pool now and stores it (409 `empty_pool` when nothing qualifies).",
        "operationId": "create_session_api_v1_courses__course_id__review_sessions_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Session",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/learning-items/{item_id}/reviews": {
      "get": {
        "operationId": "list_reviews_api_v1_learning_items__item_id__reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewRead"
                  },
                  "title": "Response List Reviews Api V1 Learning Items  Item Id  Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Reviews",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/review-sessions/{session_id}": {
      "get": {
        "operationId": "get_session_api_v1_review_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/review-sessions/{session_id}/answers": {
      "post": {
        "description": "Stores the answer, evaluates it, resolves the outcome and applies it. Always 201 once\nthe answer is stored, even if evaluation failed (`needs_self_grade`).",
        "operationId": "submit_answer_api_v1_review_sessions__session_id__answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Answer",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/review-sessions/{session_id}/end": {
      "post": {
        "operationId": "end_session_api_v1_review_sessions__session_id__end_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "End Session",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/review-sessions/{session_id}/hint": {
      "post": {
        "description": "Reveals the hint for the current question. Stored before it is returned: the answer to\nthis question earns half XP, whatever the client does next. 409 `hint_unavailable` when the\nquestion has no hint (nothing is recorded then).",
        "operationId": "reveal_hint_api_v1_review_sessions__session_id__hint_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HintState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reveal Hint",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/review-sessions/{session_id}/next": {
      "get": {
        "description": "The current question; the same one until it has an outcome or is skipped.",
        "operationId": "next_card_api_v1_review_sessions__session_id__next_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Next Card",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/review-sessions/{session_id}/skip": {
      "post": {
        "operationId": "skip_api_v1_review_sessions__session_id__skip_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Skip",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}